backend/requirements-minimal.txt
# Minimal requirements for backend startup

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9

# Core utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Basic auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Process monitoring
psutil==5.9.6

# Logging
structlog==23.2.0

# Voice Processing
elevenlabs>=0.2.0
openai-whisper>=20231117
pydub>=0.25.0

# WebSocket support
python-socketio==5.10.0
websockets==12.0